Converting majority of files to tsx
