+++ src/components/footer/footerNav.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const FooterNav = styled.div`
+    margin-left: auto;
+    margin-right: 3rem;
+    height: 60px;
+    max-width: 960px;
+  `;
+
+  export default FooterNav;
