+++ src/components/footer/footerWrapper.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const FooterWrapper = styled.footer`
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    margin: 0 auto;
+    display: block;
+    width: 100%;
+    z-index: 1000;
+    background-color: #25303b;
+}`;
+
+export default FooterWrapper;
