+++ src/components/header/headerLink.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+import { Link } from 'gatsby';
+
+const HeaderLink = styled(Link)`
+  position: relative;
+  text-decoration: none;
+  display: flex;
+  align-items: center;
+  color: #fff;
+  border: 0;
+  margin: 0;
+  margin-right: 0.5rem;
+  padding-left: 20px;
+  padding-right: 20px;
+  min-width: 42px;
+  z-index: 10;
+`;
+
+export default HeaderLink;
