+++ src/components/header/headerLinkGroup.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const HeaderLinkGroup = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
+
+export default HeaderLinkGroup;
