+++ src/components/header/headerNav.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+const HeaderNav = styled.nav`
+  margin-left: auto;
+  margin-right: auto;
+  height: 60px;
+  display: flex;
+  flex-direction: row;
+  max-width: 960px;
+  z-index: 1000;
+  justify-content: space-between;
+  overflow-x: auto;
+  overflow-y: hidden;
+  background-color: #25303b;
+`;
+
+export default HeaderNav;
