+++ src/components/header/headerWrapper.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const HeaderWrapper = styled.header`
+  position: fixed;
+  top: 0;
+  left: 0;
+  margin: 0 auto;
+  display: block;
+  width: 100%;
+  z-index: 1000;
+  background-color: #25303b;
+`;
+
+export default HeaderWrapper;
