+++ src/components/hero/hero.js
@@ -2,31 +2,9 @@ import React, { Component } from 'react';
 import { withPrefix } from 'gatsby';
 import { siteCover } from '../../../data/siteConfig';
 import styled from 'styled-components';
-
-const HeroContainer = styled.div`
-  position: relative;
-  display: table;
-  width: 100%;
-  overflow: hidden;
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: cover;
-`;
-
-const TitleContainer = styled.div`
-  display: table-cell;
-  vertical-align: middle;
-  text-align: center;
-  width: 100%;
-`;
-
-const HeroTitle = styled.h1`
-  font-weight: 700;
-  font-size: 3rem;
-  margin: 10px 60px;
-  color: #fff;
-  text-shadow: 1px 1px 4px rgba(34, 34, 34, 0.6);
-`;
+import HeroContainer from './heroContainer';
+import HeroTitle from './heroTitle';
+import TitleContainer from './titleContainer';
 
 class Hero extends Component {
   render() {
