+++ src/components/hero/heroContainer.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const HeroContainer = styled.div`
+  position: relative;
+  display: table;
+  width: 100%;
+  overflow: hidden;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: cover;
+`;
+
+export default HeroContainer;
