+++ src/components/hero/heroTitle.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const HeroTitle = styled.h1`
+  font-weight: 700;
+  font-size: 3rem;
+  margin: 10px 60px;
+  color: #fff;
+  text-shadow: 1px 1px 4px rgba(34, 34, 34, 0.6);
+`;
+
+export default HeroTitle;
