+++ src/components/hero/titleContainer.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const TitleContainer = styled.div`
+  display: table-cell;
+  vertical-align: middle;
+  text-align: center;
+  width: 100%;
+`;
+
+export default TitleContainer;
